Rough Gate Decomposition algorithm
In my write up about the Solvay-Kitaev Theorem I attempt to find my own way to decompose an arbitrary unitary get into a set of unitary matricies, I come up with an algorithm which is described above in the code. I did not prove the convergence of the algorithm and thus for basic tests it failed to decompose the unitary. In the code I considered a gate set of G = {H,X} where H is the hadamard and X is the pauli gate although the set is not necessarily universal, it should have allowed for the decomposition of Z = HXH which did not occur in the code.